chore: Move `locks/` to `application/locks/`
The source files contained in this directory have to do with managing
access rights to data resources in the context of a multi-threaded
application. As such, they belong in the category "application". This
commit reflects that containment in the directory structure.
