+++ neptune-core/src/application/database/storage/storage_schema/dbtsingleton.rs
@@ -8,7 +8,7 @@ use super::dbtsingleton_private::DbtSingletonPrivate;
 use super::traits::*;
 use super::PendingWrites;
 use super::SimpleRustyReader;
-use crate::locks::tokio::AtomicRw;
+use crate::application::locks::tokio::AtomicRw;
 
 /// Singleton type created by [`super::DbtSchema`]
 ///
