+++ neptune-core/src/application/database/storage/storage_schema/dbtsingleton_private.rs
@@ -8,7 +8,7 @@ use super::PendingWrites;
 use super::RustyValue;
 use super::SimpleRustyReader;
 use super::WriteOperation;
-use crate::locks::tokio::AtomicRw;
+use crate::application::locks::tokio::AtomicRw;
 
 // note: no locking is required in `DbtSingletonPrivate` because locking
 // is performed in the `DbtSingleton` public wrapper.
