+++ neptune-core/src/application/locks/std/atomic_rw.rs
@@ -1,25 +1,26 @@
 use std::ops::Deref;
 use std::ops::DerefMut;
 use std::sync::Arc;
-use std::sync::Mutex;
-use std::sync::MutexGuard;
+use std::sync::RwLock;
+use std::sync::RwLockReadGuard;
+use std::sync::RwLockWriteGuard;
 
+use super::shared::LockAcquisition;
 use super::traits::Atomic;
-use super::LockAcquisition;
 use super::LockCallbackFn;
 use super::LockCallbackInfo;
 use super::LockEvent;
 use super::LockType;
 
-/// An `Arc<Mutex<T>>` wrapper to make data thread-safe and easy to work with.
+/// An `Arc<RwLock<T>>` wrapper to make data thread-safe and easy to work with.
 ///
 /// # Example
 /// ```
-/// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
+/// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
 /// struct Car {
 ///     year: u16,
-/// }
-/// let mut atomic_car = AtomicMutex::from(Car{year: 2016});
+/// };
+/// let mut atomic_car = AtomicRw::from(Car{year: 2016});
 /// atomic_car.lock(|c| println!("year: {}", c.year));
 /// atomic_car.lock_mut(|mut c| c.year = 2023);
 /// ```
@@ -30,10 +31,10 @@ use super::LockType;
 ///
 /// # Examples
 /// ```
-/// # use neptune_cash::locks::std::{AtomicMutex, LockEvent, LockCallbackFn};
+/// # use neptune_cash::application::locks::std::{AtomicRw, LockEvent, LockCallbackFn};
 /// struct Car {
 ///     year: u16,
-/// }
+/// };
 ///
 /// pub fn log_lock_event(lock_event: LockEvent) {
 ///     let (event, info, acquisition) =
@@ -55,71 +56,71 @@ use super::LockType;
 /// }
 /// const LOG_LOCK_EVENT_CB: LockCallbackFn = log_lock_event;
 ///
-/// let mut atomic_car = AtomicMutex::<Car>::from((Car{year: 2016}, Some("car"), Some(LOG_LOCK_EVENT_CB)));
+/// let mut atomic_car = AtomicRw::<Car>::from((Car{year: 2016}, Some("car"), Some(LOG_LOCK_EVENT_CB)));
 /// atomic_car.lock(|c| {println!("year: {}", c.year)});
 /// atomic_car.lock_mut(|mut c| {c.year = 2023});
 /// ```
 ///
 /// results in:
 /// ```text
-/// TryAcquire lock `car` of type `Mutex` for `Read` by
+/// TryAcquire lock `car` of type `RwLock` for `Read` by
 ///     |-- thread main, `ThreadId(1)`
-/// Acquire lock `car` of type `Mutex` for `Read` by
+/// Acquire lock `car` of type `RwLock` for `Read` by
 ///     |-- thread main, `ThreadId(1)`
 /// year: 2016
-/// Release lock `car` of type `Mutex` for `Read` by
+/// Release lock `car` of type `RwLock` for `Read` by
 ///     |-- thread main, `ThreadId(1)`
-/// TryAcquire lock `car` of type `Mutex` for `Write` by
+/// TryAcquire lock `car` of type `RwLock` for `Write` by
 ///     |-- thread main, `ThreadId(1)`
-/// Acquire lock `car` of type `Mutex` for `Write` by
+/// Acquire lock `car` of type `RwLock` for `Write` by
 ///     |-- thread main, `ThreadId(1)`
-/// Release lock `car` of type `Mutex` for `Write` by
+/// Release lock `car` of type `RwLock` for `Write` by
 ///     |-- thread main, `ThreadId(1)`
 /// ```
 #[derive(Debug)]
-pub struct AtomicMutex<T> {
-    inner: Arc<Mutex<T>>,
+pub struct AtomicRw<T> {
+    inner: Arc<RwLock<T>>,
     lock_callback_info: LockCallbackInfo,
 }
 
-impl<T: Default> Default for AtomicMutex<T> {
+impl<T: Default> Default for AtomicRw<T> {
     fn default() -> Self {
         Self {
             inner: Default::default(),
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, None, None),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, None, None),
         }
     }
 }
 
-impl<T> From<T> for AtomicMutex<T> {
+impl<T> From<T> for AtomicRw<T> {
     #[inline]
     fn from(t: T) -> Self {
         Self {
-            inner: Arc::new(Mutex::new(t)),
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, None, None),
+            inner: Arc::new(RwLock::new(t)),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, None, None),
         }
     }
 }
-impl<T> From<(T, Option<String>, Option<LockCallbackFn>)> for AtomicMutex<T> {
+impl<T> From<(T, Option<String>, Option<LockCallbackFn>)> for AtomicRw<T> {
     /// Create from an optional name and an optional callback function, which
-    /// is called when a lock event occurs.
+    /// is called when a lock is event occurs.
     #[inline]
     fn from(v: (T, Option<String>, Option<LockCallbackFn>)) -> Self {
         Self {
-            inner: Arc::new(Mutex::new(v.0)),
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, v.1, v.2),
+            inner: Arc::new(RwLock::new(v.0)),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, v.1, v.2),
         }
     }
 }
-impl<T> From<(T, Option<&str>, Option<LockCallbackFn>)> for AtomicMutex<T> {
+impl<T> From<(T, Option<&str>, Option<LockCallbackFn>)> for AtomicRw<T> {
     /// Create from a name ref and an optional callback function, which
     /// is called when a lock event occurs.
     #[inline]
     fn from(v: (T, Option<&str>, Option<LockCallbackFn>)) -> Self {
         Self {
-            inner: Arc::new(Mutex::new(v.0)),
+            inner: Arc::new(RwLock::new(v.0)),
             lock_callback_info: LockCallbackInfo::new(
-                LockType::Mutex,
+                LockType::RwLock,
                 v.1.map(|s| s.to_owned()),
                 v.2,
             ),
@@ -127,7 +128,7 @@ impl<T> From<(T, Option<&str>, Option<LockCallbackFn>)> for AtomicMutex<T> {
     }
 }
 
-impl<T> Clone for AtomicMutex<T> {
+impl<T> Clone for AtomicRw<T> {
     fn clone(&self) -> Self {
         Self {
             lock_callback_info: self.lock_callback_info.clone(),
@@ -136,127 +137,110 @@ impl<T> Clone for AtomicMutex<T> {
     }
 }
 
-impl<T> From<Mutex<T>> for AtomicMutex<T> {
+impl<T> From<RwLock<T>> for AtomicRw<T> {
     #[inline]
-    fn from(t: Mutex<T>) -> Self {
+    fn from(t: RwLock<T>) -> Self {
         Self {
             inner: Arc::new(t),
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, None, None),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, None, None),
         }
     }
 }
-impl<T> From<(Mutex<T>, Option<String>, Option<LockCallbackFn>)> for AtomicMutex<T> {
-    /// Create from an `Mutex<T>` plus an optional name
+impl<T> From<(RwLock<T>, Option<String>, Option<LockCallbackFn>)> for AtomicRw<T> {
+    /// Create from an `RwLock<T>` plus an optional name
     /// and an optional callback function, which is called
     /// when a lock event occurs.
     #[inline]
-    fn from(v: (Mutex<T>, Option<String>, Option<LockCallbackFn>)) -> Self {
+    fn from(v: (RwLock<T>, Option<String>, Option<LockCallbackFn>)) -> Self {
         Self {
             inner: Arc::new(v.0),
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, v.1, v.2),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, v.1, v.2),
         }
     }
 }
 
-impl<T> TryFrom<AtomicMutex<T>> for Mutex<T> {
-    type Error = Arc<Mutex<T>>;
-    fn try_from(t: AtomicMutex<T>) -> Result<Mutex<T>, Self::Error> {
-        Arc::<Mutex<T>>::try_unwrap(t.inner)
+impl<T> TryFrom<AtomicRw<T>> for RwLock<T> {
+    type Error = Arc<RwLock<T>>;
+    fn try_from(t: AtomicRw<T>) -> Result<RwLock<T>, Self::Error> {
+        Arc::<RwLock<T>>::try_unwrap(t.inner)
     }
 }
 
-impl<T> From<Arc<Mutex<T>>> for AtomicMutex<T> {
+impl<T> From<Arc<RwLock<T>>> for AtomicRw<T> {
     #[inline]
-    fn from(t: Arc<Mutex<T>>) -> Self {
+    fn from(t: Arc<RwLock<T>>) -> Self {
         Self {
             inner: t,
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, None, None),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, None, None),
         }
     }
 }
-impl<T> From<(Arc<Mutex<T>>, Option<String>, Option<LockCallbackFn>)> for AtomicMutex<T> {
-    /// Create from an `Arc<Mutex<T>>` plus an optional name and
+impl<T> From<(Arc<RwLock<T>>, Option<String>, Option<LockCallbackFn>)> for AtomicRw<T> {
+    /// Create from an `Arc<RwLock<T>>` plus an optional name and
     /// an optional callback function, which is called when a lock
     /// event occurs.
     #[inline]
-    fn from(v: (Arc<Mutex<T>>, Option<String>, Option<LockCallbackFn>)) -> Self {
+    fn from(v: (Arc<RwLock<T>>, Option<String>, Option<LockCallbackFn>)) -> Self {
         Self {
             inner: v.0,
-            lock_callback_info: LockCallbackInfo::new(LockType::Mutex, v.1, v.2),
+            lock_callback_info: LockCallbackInfo::new(LockType::RwLock, v.1, v.2),
         }
     }
 }
 
-impl<T> From<AtomicMutex<T>> for Arc<Mutex<T>> {
+impl<T> From<AtomicRw<T>> for Arc<RwLock<T>> {
     #[inline]
-    fn from(t: AtomicMutex<T>) -> Self {
+    fn from(t: AtomicRw<T>) -> Self {
         t.inner
     }
 }
 
 // note: we impl the Atomic trait methods here also so they
 // can be used without caller having to use the trait.
-impl<T> AtomicMutex<T> {
-    pub const fn const_new(
-        t: Arc<Mutex<T>>,
-        name: Option<String>,
-        lock_callback_fn: Option<LockCallbackFn>,
-    ) -> Self {
-        Self {
-            inner: t,
-            lock_callback_info: LockCallbackInfo {
-                lock_info_owned: super::shared::LockInfoOwned {
-                    name,
-                    lock_type: LockType::Mutex,
-                },
-                lock_callback_fn,
-            },
-        }
-    }
-
-    /// Acquire read lock and return an `AtomicMutexGuard`
+impl<T> AtomicRw<T> {
+    /// Acquire read lock and return an `RwLockReadGuard`
     ///
     /// # Examples
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
     /// struct Car {
     ///     year: u16,
-    /// }
-    /// let atomic_car = AtomicMutex::from(Car{year: 2016});
+    /// };
+    /// let atomic_car = AtomicRw::from(Car{year: 2016});
     /// let year = atomic_car.lock_guard().year;
     /// ```
-    pub fn lock_guard(&self) -> AtomicMutexGuard<'_, T> {
+    pub fn lock_guard(&self) -> AtomicRwReadGuard<'_, T> {
         self.try_acquire_read_cb();
-        let guard = self.inner.lock().expect("Read lock should succeed");
-        AtomicMutexGuard::new(guard, &self.lock_callback_info, LockAcquisition::Read)
+        let guard = self.inner.read().expect("Read lock should succeed");
+        AtomicRwReadGuard::new(guard, &self.lock_callback_info)
     }
 
-    /// Acquire write lock and return an `AtomicMutexGuard`
+    /// Acquire write lock and return an `RwLockWriteGuard`
     ///
     /// # Examples
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
     /// struct Car {
     ///     year: u16,
-    /// }
-    /// let mut atomic_car = AtomicMutex::from(Car{year: 2016});
+    /// };
+    /// let mut atomic_car = AtomicRw::from(Car{year: 2016});
     /// atomic_car.lock_guard_mut().year = 2022;
     /// ```
-    pub fn lock_guard_mut(&mut self) -> AtomicMutexGuard<'_, T> {
+    pub fn lock_guard_mut(&mut self) -> AtomicRwWriteGuard<'_, T> {
         self.try_acquire_write_cb();
-        let guard = self.inner.lock().expect("Write lock should succeed");
-        AtomicMutexGuard::new(guard, &self.lock_callback_info, LockAcquisition::Write)
+        let guard = self.inner.write().expect("Write lock should succeed");
+        AtomicRwWriteGuard::new(guard, &self.lock_callback_info)
     }
 
     /// Immutably access the data of type `T` in a closure and possibly return a result of type `R`
     ///
     /// # Examples
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
     /// struct Car {
     ///     year: u16,
-    /// }
-    /// let atomic_car = AtomicMutex::from(Car{year: 2016});
+    /// };
+    /// let atomic_car = AtomicRw::from(Car{year: 2016});
     /// atomic_car.lock(|c| println!("year: {}", c.year));
     /// let year = atomic_car.lock(|c| c.year);
     /// ```
@@ -265,9 +249,8 @@ impl<T> AtomicMutex<T> {
         F: FnOnce(&T) -> R,
     {
         self.try_acquire_read_cb();
-        let guard = self.inner.lock().expect("Read lock should succeed");
-        let my_guard =
-            AtomicMutexGuard::new(guard, &self.lock_callback_info, LockAcquisition::Read);
+        let guard = self.inner.read().expect("Read lock should succeed");
+        let my_guard = AtomicRwReadGuard::new(guard, &self.lock_callback_info);
         f(&my_guard)
     }
 
@@ -275,11 +258,11 @@ impl<T> AtomicMutex<T> {
     ///
     /// # Examples
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
     /// struct Car {
     ///     year: u16,
-    /// }
-    /// let mut atomic_car = AtomicMutex::from(Car{year: 2016});
+    /// };
+    /// let mut atomic_car = AtomicRw::from(Car{year: 2016});
     /// atomic_car.lock_mut(|mut c| {c.year = 2022});
     /// let year = atomic_car.lock_mut(|mut c| {c.year = 2023; c.year});
     /// ```
@@ -288,9 +271,8 @@ impl<T> AtomicMutex<T> {
         F: FnOnce(&mut T) -> R,
     {
         self.try_acquire_write_cb();
-        let guard = self.inner.lock().expect("Write lock should succeed");
-        let mut my_guard =
-            AtomicMutexGuard::new(guard, &self.lock_callback_info, LockAcquisition::Write);
+        let guard = self.inner.write().expect("Write lock should succeed");
+        let mut my_guard = AtomicRwWriteGuard::new(guard, &self.lock_callback_info);
         f(&mut my_guard)
     }
 
@@ -298,8 +280,8 @@ impl<T> AtomicMutex<T> {
     ///
     /// # Example
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
-    /// let atomic_u64 = AtomicMutex::from(25u64);
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
+    /// let atomic_u64 = AtomicRw::from(25u64);
     /// let age = atomic_u64.get();
     /// ```
     #[inline]
@@ -314,8 +296,8 @@ impl<T> AtomicMutex<T> {
     ///
     /// # Example
     /// ```
-    /// # use neptune_cash::locks::std::{AtomicMutex, traits::*};
-    /// let mut atomic_bool = AtomicMutex::from(false);
+    /// # use neptune_cash::application::locks::std::{AtomicRw, traits::*};
+    /// let mut atomic_bool = AtomicRw::from(false);
     /// atomic_bool.set(true);
     /// ```
     #[inline]
@@ -351,13 +333,13 @@ impl<T> AtomicMutex<T> {
     }
 }
 
-impl<T> Atomic<T> for AtomicMutex<T> {
+impl<T> Atomic<T> for AtomicRw<T> {
     #[inline]
     fn lock<R, F>(&self, f: F) -> R
     where
         F: FnOnce(&T) -> R,
     {
-        AtomicMutex::<T>::lock(self, f)
+        AtomicRw::<T>::lock(self, f)
     }
 
     #[inline]
@@ -365,59 +347,95 @@ impl<T> Atomic<T> for AtomicMutex<T> {
     where
         F: FnOnce(&mut T) -> R,
     {
-        AtomicMutex::<T>::lock_mut(self, f)
+        AtomicRw::<T>::lock_mut(self, f)
     }
 }
 
-/// A wrapper for [MutexGuard] that can optionally call a callback to notify
-/// when the lock event occurs
+/// A wrapper for [RwLockReadGuard] that can optionally call a callback to
+/// notify when a lock event occurs.
 #[derive(Debug)]
-pub struct AtomicMutexGuard<'a, T> {
-    guard: MutexGuard<'a, T>,
+pub struct AtomicRwReadGuard<'a, T> {
+    guard: RwLockReadGuard<'a, T>,
     lock_callback_info: &'a LockCallbackInfo,
-    acquisition: LockAcquisition,
 }
 
-impl<'a, T> AtomicMutexGuard<'a, T> {
-    fn new(
-        guard: MutexGuard<'a, T>,
-        lock_callback_info: &'a LockCallbackInfo,
-        acquisition: LockAcquisition,
-    ) -> Self {
+impl<'a, T> AtomicRwReadGuard<'a, T> {
+    fn new(guard: RwLockReadGuard<'a, T>, lock_callback_info: &'a LockCallbackInfo) -> Self {
         if let Some(cb) = lock_callback_info.lock_callback_fn {
             cb(LockEvent::Acquire {
                 info: lock_callback_info.lock_info_owned.as_lock_info(),
-                acquisition,
+                acquisition: LockAcquisition::Read,
             });
         }
         Self {
             guard,
             lock_callback_info,
-            acquisition,
         }
     }
 }
 
-impl<T> Drop for AtomicMutexGuard<'_, T> {
+impl<T> Drop for AtomicRwReadGuard<'_, T> {
     fn drop(&mut self) {
         let lock_callback_info = self.lock_callback_info;
         if let Some(cb) = lock_callback_info.lock_callback_fn {
             cb(LockEvent::Release {
                 info: lock_callback_info.lock_info_owned.as_lock_info(),
-                acquisition: self.acquisition,
+                acquisition: LockAcquisition::Read,
             });
         }
     }
 }
 
-impl<T> Deref for AtomicMutexGuard<'_, T> {
+impl<T> Deref for AtomicRwReadGuard<'_, T> {
     type Target = T;
     fn deref(&self) -> &Self::Target {
         &self.guard
     }
 }
 
-impl<T> DerefMut for AtomicMutexGuard<'_, T> {
+/// A wrapper for [RwLockWriteGuard] that can optionally call a callback to
+/// notify when a lock event occurs.
+#[derive(Debug)]
+pub struct AtomicRwWriteGuard<'a, T> {
+    guard: RwLockWriteGuard<'a, T>,
+    lock_callback_info: &'a LockCallbackInfo,
+}
+
+impl<'a, T> AtomicRwWriteGuard<'a, T> {
+    fn new(guard: RwLockWriteGuard<'a, T>, lock_callback_info: &'a LockCallbackInfo) -> Self {
+        if let Some(cb) = lock_callback_info.lock_callback_fn {
+            cb(LockEvent::Acquire {
+                info: lock_callback_info.lock_info_owned.as_lock_info(),
+                acquisition: LockAcquisition::Write,
+            });
+        }
+        Self {
+            guard,
+            lock_callback_info,
+        }
+    }
+}
+
+impl<T> Drop for AtomicRwWriteGuard<'_, T> {
+    fn drop(&mut self) {
+        let lock_callback_info = self.lock_callback_info;
+        if let Some(cb) = lock_callback_info.lock_callback_fn {
+            cb(LockEvent::Release {
+                info: lock_callback_info.lock_info_owned.as_lock_info(),
+                acquisition: LockAcquisition::Write,
+            });
+        }
+    }
+}
+
+impl<T> Deref for AtomicRwWriteGuard<'_, T> {
+    type Target = T;
+    fn deref(&self) -> &Self::Target {
+        &self.guard
+    }
+}
+
+impl<T> DerefMut for AtomicRwWriteGuard<'_, T> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.guard
     }
@@ -428,12 +446,12 @@ impl<T> DerefMut for AtomicMutexGuard<'_, T> {
 mod tests {
     use super::*;
 
-    /// Verify (compile-time) that AtomicMutex::lock() and ::lock_mut() accept
+    /// Verify (compile-time) that AtomicRw::lock() and ::lock_mut() accept
     /// mutable values. (FnMut)
     #[test]
     fn mutable_assignment() {
         let name = "Jim".to_string();
-        let mut atomic_name = AtomicMutex::from(name);
+        let mut atomic_name = AtomicRw::from(name);
 
         let mut new_name = String::new();
         atomic_name.lock(|n| new_name = n.to_string());
