+++ neptune-core/src/application/mod.rs
@@ -1,4 +1,5 @@
 pub mod config;
 pub mod database;
 pub mod job_queue;
+pub mod locks;
 pub mod triton_vm_job_queue;
