+++ neptune-core/src/state/mod.rs
@@ -65,9 +65,9 @@ use crate::application::config::tx_upgrade_filter::TxUpgradeFilter;
 use crate::application::database::storage::storage_schema::traits::StorageWriter as SW;
 use crate::application::database::storage::storage_vec::traits::*;
 use crate::application::database::storage::storage_vec::Index;
-use crate::locks::tokio as sync_tokio;
-use crate::locks::tokio::AtomicRwReadGuard;
-use crate::locks::tokio::AtomicRwWriteGuard;
+use crate::application::locks::tokio as sync_tokio;
+use crate::application::locks::tokio::AtomicRwReadGuard;
+use crate::application::locks::tokio::AtomicRwWriteGuard;
 use crate::main_loop::proof_upgrader::ProofCollectionToSingleProof;
 use crate::main_loop::proof_upgrader::UpdateMutatorSetDataJob;
 use crate::main_loop::proof_upgrader::SEARCH_DEPTH_FOR_BLOCKS_FOR_MS_UPDATE;
